fix(anyharness): harden rung-2 per adversarial review — distinct launcher backup sibling, expectedBinary traversal guard, sha256 schema pattern, line-ceiling splits
