+++ catalogs/agents/registry.schema.json
@@ -390,7 +390,7 @@
                   },
                   "sha256": {
                     "type": "string",
-                    "minLength": 1
+                    "pattern": "^[0-9a-f]{64}$"
                   },
                   "expectedBinary": {
                     "type": "string",
