+++ scripts/max_lines_allowlist.txt
@@ -82,7 +82,6 @@ anyharness/crates/anyharness-lib/src/domains/loops/runtime.rs 802 loops runtime 
 anyharness/crates/anyharness-lib/src/live/sessions/actor/startup.rs 645 session startup (+goal/loop capability parsing, roster reconcile-on-attach); R3: +1 line initializing the new pending_stop_response field on SessionActor construction; observability PR9: +2 session.established record fields
 anyharness/crates/anyharness-lib/src/live/sessions/actor/tests/workspace_stop.rs 841 R3 real-process stop suite; every case owns a real agent process and a real ACP duplex, so the per-case setup is genuinely irreducible. +130: the dropped-stop re-entry regression (a stop future abandoned mid-escalation must leave the session answerable)
 server/tests/integration/test_organizations_api.py 624 added admin-only coverage for GET .../invitations
-anyharness/crates/anyharness-lib/src/domains/agents/model.rs 605 agents domain type barrel, previously exactly at the cap (+ResolvedAgent.credentials_from_route, the route-vs-native readiness provenance flag); split into model/ on next growth per guides/domains.md
 server/proliferate/db/store/organizations.py 777 D1 adds get_default_organization_for_user (earliest active membership = signup placement org, the v1 payer law); SG09 review restores the collision-ladder and race-backstop rationale; SSO JIT membership persistence is relocated from auth_sso into its Organization store owner; split pending with the org-store taxonomy pass
 server/proliferate/server/cloud/agent_gateway/enrollment.py 636 C-train composition re-attaches C-1 both-surface sync-completion pokes; D train: org-only ensure_signup_enrollment + per-(org,member) LiteLLM identity + fingerprint-driven per-key re-mint; shrank again when ensure_user_enrollment died (D-3), split pending
 server/proliferate/server/cloud/materialization/materialize/agent_auth.py 725 C-train composition: C-1 ensure-on-switch + C-2 delivery-ack stamping join the cut; D3 python arm adds the provider_config render branch + per-(harness,config_kind) translation table (_translate_provider_config_env) + typed-vault loading in _load_state_inputs; review fix cycle 1 hardens _hostname_first_label's userinfo/port stripping; review fix cycle 2 repairs the dangling Foundry-mapping docstring; integration adds A4 fail-closed + B3 per-harness key map on the same module D1 fail-closed comments: the renderer withholding predicate now names unfunded subjects explicitly (D train)
